requirements.txt
asgiref==3.7.2

certifi==2023.5.7

charset-normalizer==3.1.0

decouple==0.0.7

Django==4.2.1

djangorestframework==3.14.0

mysqlclient

environ==1.0

idna==3.4

pdfkit==1.0.0

pytz==2023.3

requests==2.31.0

sqlparse==0.4.4

stripe==5.4.0

urllib3==2.0.2

pymysql

whitenoise